process pool didn't work
not sure why, but this version is based on Ruby 3.1.4, it looks like it
works damn near every time, it doesn't crash until the end, and it's in
perfect shape to be shipped as the cronjob right now
