+++ Procfile
@@ -1,6 +1,6 @@
 sample: bundle exec ruby ./cli sample
 # projer: bundle exec ruby ./cli proj
-#leafer: bundle exec ruby ./cli leaf
+leafer: bundle exec ruby ./cli leaf
 # leafer2: bundle exec ruby ./cli leaf2
 # leafer3: bundle exec ruby ./cli leaf3
 # leafer4: bundle exec ruby ./cli leaf4
